refactor: useUserを共通のhooksに変更
